+++ edx_django_utils/monitoring/scripts/new_relic_nrql_search.py
@@ -1,17 +1,10 @@
 """
-This script finds deprecated uses of 'code_owner', to aid in the transition to 'code_owner_squad'.
+This script takes a regex to search through the NRQL of New Relic alert policies
+and New dashboards.
 
-Sample usage::
+For help::
 
-    python edx_django_utils/monitoring/scripts/new_relic_code_owner_audit.py
-
-    python edx_django_utils/monitoring/scripts/new_relic_code_owner_audit.py --policy_id 5 --policy_id 17
-
-    python edx_django_utils/monitoring/scripts/new_relic_code_owner_audit.py --dashboard_id 23
-
-For more help::
-
-    python edx_django_utils/monitoring/scripts/new_relic_code_owner_audit.py --help
+    python edx_django_utils/monitoring/scripts/new_relic_nrql_search.py --help
 
 """
 import os
@@ -20,42 +13,64 @@ import re
 import click
 import requests
 
-# Finds deprecated uses of 'code_owner'. Using 'code_owner' in a FACET or IS NULL check is acceptable.
-CODE_OWNER_REGEX = re.compile(r'code_owner\s*(=|!=|LIKE|NOT LIKE|RLIKE|NOT RLIKE|IN)')
-
 
 @click.command()
+@click.option(
+    '--regex',
+    required=True,
+    help="The regex to use to search NRQL in alert policies and dashboards.",
+)
 @click.option(
     '--policy_id',
     type=int,
     multiple=True,
-    help="A specific policy id to check. Multiple can be supplied.",
+    help="Optionally provide a specific policy id to check. Multiple can be supplied.",
 )
 @click.option(
     '--dashboard_id',
     type=int,
     multiple=True,
-    help="A specific dashboard id to check. Multiple can be supplied.",
+    help="Optionally provide a specific dashboard id to check. Multiple can be supplied.",
 )
-def main(policy_id, dashboard_id):
+def main(regex, policy_id, dashboard_id):
     """
-    Search New Relic alert policies for uses of 'code_owner'.
+    Search NRQL in New Relic alert policies and dashboards using regex.
+
+    Example usage:
+
+        new_relic_nrql_search.py --regex tnl
+
+    Note: The search ignores case since NRQL is case insensitive.
+
+    Pre-requisite, set the following environment variable (in a safe way):
+
+    NEW_RELIC_API_KEY=XXXXXXX
+
+    See https://docs.newrelic.com/docs/apis/intro-apis/new-relic-api-keys/#user-api-key for details
+    on setting an API key.
+
+    To skip alert policies or dashboards, just use a non-existent id, like --policy_id 0 or --dashboard_id 0.
+
+    WARNING: NRQL Baseline alert conditions can't be searched. See
+    https://docs.newrelic.com/docs/alerts-applied-intelligence/new-relic-alerts/rest-api-alerts/rest-api-calls-alerts/#excluded
     """
     # Set environment variables
     api_key = os.environ['NEW_RELIC_API_KEY']
     headers = {"X-Api-Key": api_key}
+    compiled_regex = re.compile(regex)
 
-    audit_alert_policies(headers, policy_id)
+    audit_alert_policies(compiled_regex, headers, policy_id)
     print()
-    audit_dashboards(dashboard_id, headers)
+    audit_dashboards(compiled_regex, dashboard_id, headers)
     print(flush=True)
 
 
-def audit_alert_policies(headers, policy_id):
+def audit_alert_policies(regex, headers, policy_id):
     """
-    Searches New Relic alert policies for 'code_owner'.
+    Searches New Relic alert policy NRQL using the regex argument.
 
     Arguments:
+        regex (re.Pattern): compiled regex used to compare against NRQL
         policy_id (tuple): optional tuple of policy ids supplied from the command-line.
         headers (dict): headers required to make http requests to New Relic
     """
@@ -76,7 +91,7 @@ def audit_alert_policies(headers, policy_id):
     # Note: policy_id is an optional tuple of policy ids supplied from the command-line.
     if policy_id:
         policies = [policy for policy in policies if policy['id'] in policy_id]
-    print(f"Searching for code_owner in {len(policies)} alert policies...")
+    print(f"Searching for regex {regex.pattern} in {len(policies)} alert policies...")
     policy_ids_printed = {}
     for policy in policies:
         print('.', end='', flush=True)
@@ -92,17 +107,16 @@ def audit_alert_policies(headers, policy_id):
 
         for nrql_condition in response_data['nrql_conditions']:
             nrql_query = nrql_condition['nrql']['query']
-            # check if 'code_owner' is found in the NRQL
-            if CODE_OWNER_REGEX.search(nrql_query, re.IGNORECASE):
+            if regex.search(nrql_query, re.IGNORECASE):
 
-                # Print the alert policy header for the first alert condition with 'code_owner'
+                # Print the alert policy header for the first alert condition matched
                 if policy['id'] not in policy_ids_printed:
                     policy_ids_printed[policy['id']] = True
                     print('\n')
                     print(f"Found in {policy['id']}: {policy['name']}:")
                     print('')
 
-                # Print the alert condition that contains 'code_owner'
+                # Print the alert condition that matched
                 print(f"- {nrql_condition['name']}: {nrql_query}")
 
     if policy_ids_printed:
@@ -111,7 +125,7 @@ def audit_alert_policies(headers, policy_id):
             command_line += f'--policy_id {policy_id} '
         print("\n\nRun again with found policies: {}".format(command_line))
     else:
-        print("\n\nNo alert policies found with deprecated use of 'code_owner'.")
+        print("\n\nNo alert policies matched.")
 
     print(
         "\nWARNING: NRQL Baseline alert conditions can't be searched. See "
@@ -120,11 +134,12 @@ def audit_alert_policies(headers, policy_id):
     )
 
 
-def audit_dashboards(dashboard_id, headers):
+def audit_dashboards(regex, dashboard_id, headers):
     """
-    Searches New Relic dashboard widgets for 'code_owner'.
+    Searches New Relic alert policy NRQL using the regex argument.
 
     Arguments:
+        regex (re.Pattern): compiled regex used to compare against NRQL
         dashboard_id (tuple): optional tuple of dashboard ids supplied from the command-line.
         headers (dict): headers required to make http requests to New Relic
     """
@@ -146,7 +161,7 @@ def audit_dashboards(dashboard_id, headers):
     # Note: dashboard_id is an optional tuple of dashboard ids supplied from the command-line.
     if dashboard_id:
         dashboards = [dashboard for dashboard in dashboards if dashboard['id'] in dashboard_id]
-    print(f"Searching for code_owner in {len(dashboards)} dashboards...")
+    print(f"Searching for regex {regex.pattern} in {len(dashboards)} dashboards...")
     dashboard_ids_printed = {}
     for dashboard in dashboards:
         print('.', end='', flush=True)
@@ -166,17 +181,16 @@ def audit_dashboards(dashboard_id, headers):
                     continue
 
                 nrql_query = data['nrql']
-                # check if 'code_owner' is found in the NRQL
-                if CODE_OWNER_REGEX.search(nrql_query, re.IGNORECASE):
+                if regex.search(nrql_query, re.IGNORECASE):
 
-                    # Print the dashboard header for the first widget nrql with 'code_owner'
+                    # Print the dashboard header for the first widget nrql that matches
                     if dashboard['id'] not in dashboard_ids_printed:
                         dashboard_ids_printed[dashboard['id']] = True
                         print('\n')
                         print(f"Found in {dashboard['id']}: {dashboard['title']}:")
                         print('')
 
-                    # Print the widget NRQL that contains 'code_owner'
+                    # Print the widget NRQL that matches
                     print(f"- {widget['presentation']['title']}: {nrql_query}")
 
     if dashboard_ids_printed:
@@ -185,7 +199,7 @@ def audit_dashboards(dashboard_id, headers):
             command_line += f'--dashboard_id {dashboard_id} '
         print("\n\nRun again with found dashboards: {}".format(command_line))
     else:
-        print("\n\nNo dashboards found with deprecated use of 'code_owner'.")
+        print("\n\nNo dashboards found that match.")
 
 
 if __name__ == "__main__":